create new conversations and new mssages (givne the convo id) work.  Including a back and forth where the conversation is loaded inot the messages and the system "remembers" previous messages.  I commened out a bunch of endpoints to make it easier to track which one was in operation.
No sure what some of the cahnges in basic_chat are for. the Id check cured an error where the id was being treated as a string.  I think the other changes track the 'user' vs 'assistant' roles.

Took out conversation ID fro MessageCreate class to it's not in the json payload.  Not sure if that's going to have consequences, but it seemed redundant given that the id would be in the URL>