The conversatiln list loads.
the create new converation button creates a new conversation.

BUT, the conversation list does not update with the new conversation.

Have had a very tought time getting the create new conversaiton button working.  Had a situtaiotn where the createNewCovnersation fucitons was in the ConversationContext but an infininate loop was being created.  So, I tried to simplify things but hust having the button in the conversation list itself.

I might try to move the function back out, now that I can actually hit the api and have it work. First, however, i'm going to get the conversation list to update automatically.